Added version independent Executors
Added:
- CustomCommandExecutor.java
- PlayerExecutor.java
- VPBaseExecutor.java
- VPExecutor.java
- VirtualPlayerListener.java

Previously, these were located in the nms.{version} packages/modules.
By design, it was necessary to have these classes duplicated per module
because they contained version dependent code.

The problem with having them duplicated in each module is that it was
tedious to make changes, or make additions because each module would have
to be modified & updated. But with the introduction of the VirtualPlayer
interface, they have now been made version independent.

Now, any changes or additions to these classes will be propagated to each
version of Minecraft... As long as there is an implementation module that
implements the VirtualPlayer interface and the VirtualPlayerFactory
abstraction.