VirtualPlayers v1.6.6
This is a temporary, intermediate version just before the release of v2.0
which will contain backwards-incompatible changes.

v1.6.6 adds the VirtualPlayer interface while hopefully maintaing
backwards compatibility.

This commit also contains a method generateErrorOnPurose() which is a self
reminder, in the form of a compile-time error that cannot be ignored:

The reminder is to fix VirtualPlayers.getOnlinePlayers()

First, the implementation details must change. Right now, the problem is
that it uses Bukkit.getOnlinePlayers() which returns a Collection<Player>:
This cannot be used on older versions of Minecraft. So it breaks backwards
compatibility. Planned for v1.6.7

Next, the return value should probably changed from an array of Player[]
to a Collection<Player>. Planned for v2.0